backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
Pillow==10.1.0
numpy==1.26.2
opencv-python-headless==4.8.1.78
# tensorflow==2.15.0 # Commented for simpler installation in skeleton
gunicorn==21.2.0
python-dotenv==1.0.0
pytest==7.4.3
